Make GPG key management more robust

+++ scripts/gpg-init.sh
@@ -29,6 +29,8 @@
 
 set -exuo pipefail
 
+TMP_GPG_HOME=$( mktemp -d -t 'XXXX' )
+
 PRIVATE_KEYRING=$1
 PUBLIC_KEYRING=$2
 PRIVATE_KEY="$(basename ${PRIVATE_KEYRING} .gpg).asc"
@@ -40,48 +42,55 @@ if [[ $# -lt 2 ]] ; then
 fi
 
 function gpginitcleanup {
-  rm -f ${TMP_PRIVATE_KEYRING}
-  rm -f ${TMP_PUBLIC_KEYRING}
+  [[ -f ${TMP_PRIVATE_KEYRING} ]] && rm -f ${TMP_PRIVATE_KEYRING}
+  [[ -f ${TMP_PUBLIC_KEYRING} ]] && rm -f ${TMP_PUBLIC_KEYRING}
 }
+
 function gpginitend {
     export EXIT=$?
     if [[ $EXIT != 0 ]]; then
       gpginitcleanup
     fi
+    gpginitcleanup
     exit $EXIT
 }
 
-trap gpginitend EXIT
+trap gpginitend ERR
 
-TMP_PRIVATE_KEYRING='./scripts/rkt.sec'
-TMP_PUBLIC_KEYRING='./scripts/rkt.pub'
+TMP_PRIVATE_KEYRING="./scripts/rkt.sec"
+TMP_PUBLIC_KEYRING="./scripts/rkt.pub"
 
 WORKING_DIR=$(cd -P -- "$(dirname -- "$0")" && pwd -P)
 
-if [ ! -f ${PRIVATE_KEY} ] || [ ! -f ${PRIVATE_KEYRING} ]; then
-  echo "Signing GPG secret keyring ${PRIVATE_KEYRING} or secret key ${PRIVATE_KEY} NOT found!"
+if [[ ! -f ${PRIVATE_KEY} ]]; then
+  echo "Signing GPG secret key ${PRIVATE_KEY} NOT found!"
   gpginitcleanup
-  echo "Creating GPG secret keyring ${PRIVATE_KEYRING} and secret key ${PRIVATE_KEY}."
+  echo "Creating GPG secret key ${PRIVATE_KEY}."
   pushd ${WORKING_DIR}
     pushd ..
       # Create secret and public key
       gpg --no-tty --batch --gen-key ./scripts/gpg-batch
+      #gpg --no-tty --no-default-keyring --trust-model always  --secret-keyring ${TMP_PRIVATE_KEYRING} --keyring ${TMP_PUBLIC_KEYRING} --import ${TMP_PRIVATE_KEY}
       # Amend KEY_ID selection to use --with-colon
-      #KEY_ID=`gpg --no-tty --no-default-keyring --secret-keyring ${TMP_PRIVATE_KEYRING} --import ${TMP_PRIVATE_KEYRING} --allow-secret-key-import  2>&1 | awk 'NR==1 { sub(/:/,"",$3) ; print $3 }'`
-      KEY_ID=`gpg --no-tty --no-default-keyring --keyring ${TMP_PUBLIC_KEYRING} --import ${TMP_PUBLIC_KEYRING} 2>&1 | awk 'NR==1 { sub(/:/,"",$3) ; print $3 }'`
+      KEY_ID=$(gpg --no-tty --no-default-keyring --secret-keyring ${TMP_PRIVATE_KEYRING} --keyring ${TMP_PUBLIC_KEYRING} --list-keys --with-colons|grep pub|cut -d':' -f5)
+      #KEY_ID=`gpg --no-tty --no-default-keyring --keyring ${TMP_PUBLIC_KEYRING} --import ${TMP_PUBLIC_KEYRING} 2>&1 | awk 'NR==1 { sub(/:/,"",$3) ; print $3 }'`
       echo -e "trust\n5\ny\n" | gpg --no-tty --no-default-keyring --trust-model always --command-fd 0 --keyring ${TMP_PUBLIC_KEYRING} --edit-key ${KEY_ID}
       # Export secret key as armored text
       gpg --no-tty --no-default-keyring --armor --secret-keyring ${TMP_PRIVATE_KEYRING} --keyring ${TMP_PUBLIC_KEYRING} --export-secret-key ${KEY_ID} >${PRIVATE_KEY}
       # Export public key as armored text
       gpg --no-tty --no-default-keyring --armor --secret-keyring ${TMP_PRIVATE_KEYRING} --keyring ${TMP_PUBLIC_KEYRING} --export ${KEY_ID} >${PUBLIC_KEY}
-      gpg --no-tty --no-default-keyring --armor --export-ownertrust > ownertrust-gpg.txt
+      #gpg --no-tty --no-default-keyring --armor --export-ownertrust > ownertrust-gpg.txt
       # Provide keyrings as requested
-      mv ${TMP_PRIVATE_KEYRING} ${PRIVATE_KEYRING}
-      mv ${TMP_PUBLIC_KEYRING} ${PUBLIC_KEYRING}
-      chmod 400 ${PRIVATE_KEYRING}
-      chmod 400 ${PUBLIC_KEYRING}
+      #mv ${TMP_PRIVATE_KEYRING} ${PRIVATE_KEYRING}
+      #mv ${TMP_PUBLIC_KEYRING} ${PUBLIC_KEYRING}
+      #chmod 400 ${PRIVATE_KEYRING}
+      #chmod 400 ${PUBLIC_KEYRING}
       chmod 400 ${PRIVATE_KEY}
       chmod 400 ${PUBLIC_KEY}
+      rm -f ${TMP_PRIVATE_KEYRING}
+      rm -f ${TMP_PUBLIC_KEYRING}
+      rm -f ${PRIVATE_KEYRING}
+      rm -f ${PUBLIC_KEYRING}
     popd
   popd
 fi
